Refactoring HomeView() to take HomeViewModel()
