+++ Expenses/Home/ViewModel/HomeViewModel.swift
@@ -8,19 +8,12 @@
 import Foundation
 import UIKit
 
-
-enum ActiveView: Identifiable {
-    
-    case settings
-    case add
-    case all
-    
-    var id: Int {
-        hashValue
-    }
-}
-
 class HomeViewModel: ObservableObject {
+    
+    //TODO: Have Dispatchque allocated to main thread when refactoring dataHandler for HomeView()
+    
+    let dataManager = CoreDataHandler.shared
+    
     
     
     init() {
@@ -49,6 +42,62 @@ class HomeViewModel: ObservableObject {
         }
         
         return message
+    }
+    
+    var spentToday: Double {
+        
+        var total = 0.0
+        
+        for transaction in dataManager.today {
+            total += transaction.amount
+        }
+        
+        return total
+    }
+    
+    var spentYesterday: Double {
+        
+        var total = 0.0
+        
+        if dataManager.yesterday.isEmpty {
+            total = 1.0
+        } else {
+            for transaction in dataManager.yesterday {
+                total += transaction.amount
+            }
+        }
+        return total
+    }
+    
+    var diffPercentage: Double {
+        let difference = spentToday - spentYesterday
+        if dataManager.yesterday.isEmpty && dataManager.today.isEmpty {
+            return 0.0
+        } else {
+            return (difference / spentYesterday) * 100.099
+            
+        }
+    }
+    
+    var topCat: String {
+        var arry: [String] = []
+        
+        for transaction in dataManager.all {
+            arry.append(transaction.category ?? "")
+            
+        }
+        
+        return arry.filtered().first ?? "No Category Recorded"
+    }
+    
+    var topPayment: String {
+        var arry: [String] = []
+        
+        for transaction in dataManager.all {
+            arry.append(transaction.bank ?? "")
+        }
+        
+        return arry.filtered().first ?? "No Payment Recorded"
     }
 }
